dng: persona data screen unit complete - open, arcana grid, layout and draw matched
All eight routines of ui/personadata are C and exact: PersonaDataOpen
(the screen's eight images out of the archive preloaded at 0x80180000),
ArcanaGridOpen/Step, PersonaDataLayout and PersonaDataDraw.

- tilemap.h: TILEMAP_INT_COUNT, the int-returning FormatDecimal and int
  count TileMapWriteRowRev dng's page was built against; slot.h's
  SLOT_SETPOS_INT now takes x as int too (only this unit uses it).
- sym: dng's BgReset, SoundOpenSeq, CellsInit, ImageAnimStopAll,
  MenuResetRepeat, DrawSpellName; g_bg_map0-2, g_bg_layer_otz, g_ot sized,
  g_arcana_*, g_hud_digits split out of its bss blob, the pack offsets and
  a few literal stores ignore:True.

+++ include/persona/common/slot.h
@@ -93,9 +93,9 @@ extern void SlotInitTagged(void *def, int slot, int attr, short x, short y);
 #else
 extern void SlotInitTagged(void *def, u_char slot, int attr, short x, short y);
 #endif
-/* And again: dng's persona data pick hands y over without narrowing it. */
+/* And again: dng's persona data screen hands x and y over unnarrowed. */
 #ifdef SLOT_SETPOS_INT
-extern void SlotSetPos(u_char slot, int attr, short x, int y);
+extern void SlotSetPos(u_char slot, int attr, int x, int y);
 #else
 extern void SlotSetPos(u_char slot, int attr, short x, short y);
 #endif
